Add pip.resolve
Move out code (pip.req.req_set.prepare_files and support) out into pip.resolve
